Changes to replace image tag versions
The latest image pushed to noiro registry will have
a .z tag. There is another tag associated with this z
tag that will have the date and job id. We will use this
tag to replace the released tags defined in the acc-provision
versions.yaml such that each acc-provision package pushed to
pypi can map to unique image tags which do not change (like the
.z tag).