Adding a provision to fetch latest sha in case of new manual push to release status page
While doing a release, we end up making a manual push to docker as
docker doesnt support tagging to same image sha. As such sha changes
and that is not tracked by the release page. We have to provide a way
where we make status idempotent of the case where whther it is a rollback
tagging to older sha or a push to new sha , the release page should be updated
accordingly.